Changes include: - Allow to configure the SDS stream IRI via RDF-connect config - Define RDF-Connect processor in its own source file
